Add files via upload
This is a Toolhead sensor mount for the ERCF with an Orbiter V2 use. You need a Microswitch and a 4mm steel ball which comes between microswitch and the filamentpath. Its a little bit tricky to find the right distance. But it works very well. I screw the microswitch with a m2 self tapping screw. Use screws which works for you. I reduced the size a little bit.  Therefore it looks a little bit different than in the photo.
You can also use it with a Sherpa Micro but i didnt tested jet. I woult insert below the mount two Headinserts to screw the Sherpa Micro down.